+++ app/src/screens/Dialer/BackspaceButton.tsx
@@ -7,7 +7,11 @@ const BackspaceSource = require('../../../assets/icons/backspace.png');
 export type Props = Pick<ButtonProps, 'disabled' | 'onPress'>;
 
 const BackspaceButton: React.FC<Props> = ({ disabled, onPress }) => (
-  <Button size={96} disabled={disabled} onPress={onPress}>
+  <Button
+    size={96}
+    disabled={disabled}
+    onPress={onPress}
+    testID="back_space_button">
     <Image source={BackspaceSource} resizeMode="contain" style={styles.image} />
   </Button>
 );