+++ app/src/util/__tests__/env.test.ts
@@ -1,0 +1,10 @@
+import { getEnvVariable } from '../env';
+
+describe('env variable util', () => {
+  it('should notify user if missing env', () => {
+    const testVariable = 'something random';
+    expect(() => {
+      getEnvVariable(testVariable);
+    }).toThrowError(`Missing ${testVariable} in .env`);
+  });
+});