+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * Upgrade React Native from `0.70.6` to `0.70.9`.
   * Potentially fixes iOS builds for newer versions of Xcode.
 * Added e2e tests for outgoing calls
+* Added unit tests for store and screens